[FIX] sap.m.FacetFilter: timing issue with event firing is now resolved
Problem description:
On touch support device, when user opens a facet popover
and presses directly over the control reset button,
then "reset" event is fired before "listClose" event.

Solution:
There is a close popover animation on a touch support devices.
We are waiting on touch support devices for this animation to
finish before we fire the "reset" event.

BCP: 2180101997
CR-Id: 002075125800001412292021
Change-Id: I0c730221a2bb068c6979543872ed51f4519bafa2