[INTERNAL] Demo Kit: Added 'yaml' and 'yml' extensions for files in code viewer
Problem: When inspecting files with the extensions 'yaml' or 'yml' in
the code viewer of the Demo Kit, instead of seeing the content of the
file, we see the character codes of the file's content.

Solution: We add the 'yaml' and 'yml' extensions in the list of files
which are supposed to be interpreted as non-binary.

Change-Id: I823cff627871f8c954037d88a4c167536a25a36f
BCP: 2170103637