[INTERNAL] JSDoc: introduce the @template tag for type parameters
With the @template tag, generic functions can document their type
parameters. The @template tag must have a name, can have a type (which
then represents the type parameter's constraint) and a default value.

See https://www.typescriptlang.org/docs/handbook/jsdoc-supported-types.html#template
for TypeScript's documentation of this tag.

The current implementation supports only one type parameter per tag, the
tag is only supported for functions/methods, not for classes or types
and it cannot be used for methods that are marked with @ui5-metamodel.

Change-Id: I53bb47e632bf64aeecb41633eebf5a8796e41cbf