[INTERNAL] sap.ui.core: enable custom type fragment creation
When custom fragment types are used and do not provide the async loading mechanism (via load API),
invalid fragment names are used to load them.
Therefore the asynchronous loading is only used in case of UI5 provided fragment types.
With that the fragments of a custom type are loaded synchronously.

BCP: n/a
CR-Id: 002075125800001432842021
Change-Id: Iee197e4620dbfe5f4a2ec77a26c47575d2b823e6