[FIX] ui.Table: Remove 'aria-labelledby' attribute from table cell when 'columnHeaderVisible=false'
The column header does not get announced by JAWS when 'columnHeaderVisible' is set to 'false' on the table. The reason for this is because when setting 'columnHeaderVisible' to 'false' the column header(s) get not rendered in the DOM and the focused table cell points with its 'aria-labelledby' attribute to a none existing DOM element. Aligning to the SAP ACC Standard a label must be visible on the UI. Setting the property to 'false' makes the table not accessible to 100% therefore the 'aria-labelledby' attribute should not be rendered to the table cell.

BCP: 2080298597
Change-Id: I9f5962fed0fe88913b09686ce268ec0eaa8d7511