[INTERNAL][test] sap.ui.integration: Card editor
a. set the control from text to input or MultiComboBox if the editable is false in ListField.
b. set the control from input to input or ComboBox if the editable is false in StringField.
c. add visible property for the controls in StringField and ListField.
d. add test case to check the string parameter(not defined with values) with editable false or visible false in translation mode.
e. add test cases to check the string/list parameers(defined with values) with editable false or visible false.
JIRA: SMPADMSH4-3290

Change-Id: Ib5792002608984797608238898540b2affab52c5