[INTERNAL] JSDoc: Add validation for metadata 'library' information
- Adds validation within the JSDoc build for consistent metadata 'library'
  information
- Fixes finding in sap.m.TabContainerItem

REBASE REQUIRED:
Open changes need to be rebased in order to be properly validated in the
voter.
This is because the jsdoc plugin is not built in a normal voter run.
The new validation check will be active for all openui5 changes, also for
those not based on this commit. This will cause the docs voter to fail
with an error for the TabContainerItem.

Change-Id: Ied32b67a8b18710041181f3df8714e13093c6870
JIRA: CPOUI5FOUNDATION-329