[FIX] sap.ui.rta: hideFromReveal also for deeper structures
Issue with addSection in our demoapplications. Sections
acts as container for deep structures inside. We need to check
hideFromReveal attribute just for properties that are related to
the selected element. Now additionally to the bindingPaths the
relevant container of the overlay is checked that is bound to a
property with hideFromReveal.

Change-Id: I146b1feeb20c48b224783bbabb237e458a1b2766
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
BCP: 2170030560