[INTERNAL]AdaptationFilterBar: ensure correct cleanup
The 'executeRemoves' needs to be executed by the controller as the
AdaptationFilterBar does now know when the Dialog has been closed.

In addition on exit the created FilterItems are checked for leftovers
to ensure that everything has been cleaned up on exit.

This change ensures that the FilterBar is being cleaned up correctly
on destroying the AdaptationFilterBar. Without this change the
Dialog might not open again if the user navigates back into the app
after leaving it through the FLP.

BCP: 2170080012
BCP: 2170135410
CR-Id: 002075125800001686782021
Change-Id: I675d5784794316fdf9bd767b6b8e8f2ab3e67eee
(cherry picked from commit 7ab7f111d0940f060d5ecf2e3cb670913c92583a)