[INTERNAL] ui.Table: Allow column resizing when 'columnHeaderVisible=false'
During the debugging of BCP:2080298597  I noticed that resizing a column, when columnHeaderVisible=false, was not even possible.
Because the column resize ruler gets all relevant information from the element '.sapUiTableColHdrCnt' which by default has 'display: none' it never worked.
Instead now we rely on the table content itself (.sapUiTableCCnt) that is always available.

Change-Id: Ib187eadae773135ec9742307d25093d583a2d572