[INTERNAL] sap.ui.rta: Save variant after LocalReset
If a local reset is executed on a control variant the variant gets
dirty.
As we currently don't create a dedicated change for the local reset, it
is not possibly to save the changes to a new variant.
Easiest solution for now is to always assign the changes to the current
control variant via a composite command.

Change-Id: I1755f2efd6a19ca1e99015e8c203e9545d1625ff
JIRA: TEAM42-115