[FIX] sap.m.Tokenizer: Improve NMore opening UX
When there are too many tokens and the N-More gets pressed,
we need to ensure smooth opening of the Popover.
In order to achieve that and to avoid layout thrashing, load
a set of items prior Popover's openning and when ready
load rest of the items.

BCP: 2170111397
JIRA: BGSOFUIRILA-3113

Critical Impact
The incident was raised as medium impact,
however MultiInput by documentation could have 200+ items and this is
an issue in IE. It takes too much time there to open the popover.
This change improves this. 1.87 is the latest release where IE11 is
supported.

Change-Id: Iad315489bf95493d97e7daf60cb5aaf7a3d93bc3
CR-Id: 002075125800001574672021