[INTERNAL] sap.ui.dt - aggregationBindingTemplates in DesignTime
It is now possible to attach overlays belongs to an aggregation binding
as 'aggregationBindingTemplateOverlays' aggregation in ElementOverlays.
DesignTime checks for bounded template elements and creates also
Overlays for the templates. As the template control is not striclty
related to an element, the overlays of the template control gets
attached not as children but as aggregationBindingTemplateOverlays to
the parent overlay.

Change-Id: I5747714e4f152544192b071edc09d47dd7d47ea1
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
JIRA: TEAM42-110