[INTERNAL] Integration Cards: Fix resource root when card is located in server root
When card is located in the server's root
E.g. http://localhost:80/card.json
The resource root is registered with empty url and this is causing
component cards to not work.
Now it fallbacks to "/" and component cards work.

Unit test was not possible for this use case exactly,
because of the test setup which we have.
Added test for the general use case.

Change-Id: Ifc4115acf049ab23a3ecb03a921e474cbd162c67
BCP: 2170144218