[INTERNAL] sap.ui.fl KeyUserConnector only allows one way to share variants
With the UI5 version 1.90 there should be only one way how to share variants since the public option is only available for Key Users in this setup.
As soon as the service is capable of handling the comp variants within RTA, there should be a hard switch even if later on the PUBLIC option is available for a broader user base.

Change-Id: Ib3435dc163d7256ecba5a39594cfd6651e20003c