[INTERNAL] AlignedFlowLayout: avoid themeChanged events during test run
A themeChanged event during test execution could result in unexpected
.reflow() calls.

Configuring the Core to wait for the theme to be loaded and waiting with
the test execution until the Core has initialized should avoid such
problems.

At the same time, the test was refactored to make use of the generic
test starter. The test starter enables async loading and already waits
for the Core by default.

Change-Id: I09ac0166dd4d111e04bf163faa4b382b6294a2e2