[INTERNAL] sap.m.DatePicker: aria-roledescription update
Because of DatePicker's aria-roledescription attribute,
screen readers was not announcing the combobox role.
Without knowing that, screen reader users might not be
able to use the required keys to open the calendar popup.
That's why now in the roledescription also "combobox" is added.

This change has high impact on several customers.

Change-Id: I04fb1f9404ad85cc8494801d3f72efded41755c5
CR-Id: 002075125900001611842021
BCP: 2170037211