[FIX] InputBase: retaining wrong DOM value in onAfterRendering is fixed
 - InputBase was designed to retain the input value between renderings.
 - To do this properly InputBase checks the DOM value of the input field
before and after rendering.
 - With DOM based rendering (apiVersion:2) control DOM node might be
deleted before the rendering of a control during the partial DOM update
of the parent. With string based rendering, since the output of the
parent depends the output of descendants controls, the DOM node removal
of the control was not possible during the rendering.
 - This change now ignores retaining the input value in onAfterRendering
if there is no active DOM in the onBeforeRendering otherwise input field
always shows an empty value.

BCP: 2180144530

Change-Id: I4c8844b0a6052e247aafef95709996b617098d41