[FIX] GridRenderer: Provide renderer key for control wrappers
 - To prevent different logical subtrees being reused, a unique ID must
must be provided in the RM#openStart/voidStart API. This is especially
very important when rendering over aggregations.
 - Using a key prevents DOM nodes corresponding to separate child
controls from being seen as a diff that needs to be reconciled. In this
case, for example, a newly added control at the head of aggregations
causes only a new corresponding control DOM node to be created rather
than all the aggregation DOM nodes being updated (if node type matches)
or re-created (if the node type does not match).
 - In general, for better rendering performance and to correctly
maintain the focus, the key must be provided if the parent's DOM
structure might change.

BCP: 2180144530

Change-Id: I7bb24f0e761711cfd7c3374267fd9a7f325181ea