[INTERNAL] Correctly inherit parent renderers
When a renderer defines apiVersion:2, but the control class definition
only picks the 'render' function from the renderer, then a new renderer
will be created on the fly which is missing the apiVersion property.

Instead, the whole renderer should be referenced in the control class
definition.

Also fixes inheriting renderers that didn't set their own apiVersion:2.

Change-Id: Ie083301922606914d47d72888c0cdeea72e630ac