[INTERNAL] sap.ui.fl,sap.ui.rta,sap.ui.integration: add library property
When a Control or Element class resides in a sub-package of a library
and if the name of that sub-package is all-lowercase, then the runtime
guesses the wrong library name from the control/element name.

In such cases, the correct library name has to be specified in the
runtime metadata.

Change-Id: I63b2b5cea75517a728be56f0aa0eb2ef0d47b4a3