[INTERNAL] sap.ui.core: allow testsuites to add own qunit/sinon versions
To allow library-local versions of QUnit or Sinon, testsuites must be
able to configure the location of such versions.

Basically, the hard-coded configuration for the existing versions has
been moved to the general configuration object and the starter reads the
required information from there.

Testsuites can enhance the configuration in the usual way:

  {
     qunit: {
        versions: {
           "2.x": {
              module: "<module ID of QUnit 2.x>",
              css: "<resource name of the QUnit 2.x CSS file>"
           }
        }
        version: "2.x"
     },
     sinon: {
        versions: {
           "10.x": {
              module: "<module ID of Sinon 10.x>",
              bridgeModule: "<module ID of the corresponding bridge>"
           }
        },
        version: "10.x"
     }
  }

The "versions" maps will be merged from testsuite configuration and
default configuration and "version" can refer to any version key in the
map.

Custom Sinon versions can only be used via the above configuration.
There is no way to configure dependency shims for them so that they
could be required in JavaScript code. Such shims are only configured for
the built-in Sinon versions (1 and 4).

Change-Id: Id9e0af7c92e9b2a2a94e08378267a7a44226de79