[FIX] sap.ui.unified.FileUploader: After opening dialog are now not trapped with the focus
When you press the button, a dialog opens up, so now the keyup event is fired elsewhere,
therefore the button stays in infinite space pressed state, which prevents every other
interaction. After the change, we manually take down the flag, so the button can be used normally.
There is no test(tested manually), because on space press, a browser dialog is opened and this intercept
the testing.

BCP: 2170145992

Change-Id: Iebb82ed4f1f6a2a996dca66a58ad9b0759791cac