[INTERNAL] sap.m.OverflowToolbarAssociativePopover: Rendering rule fix
Problem: OverflowToolbarAssociativePopover doesn't comply with with the
Rendering support rule and throws an error in the Support Assistant.

Root Cause: OverflowToolbarAssociativePopover doesn't have it's own
rendering file that specifies that its rendering apiVersion is 2. Even
though the OverflowToolbarAssociativePopover extends the Popover which
has the apiVersion set to 2, it doesn't have such own file and property.
This is considered an issue by the RenderManager's getApiVersion method.

Solution: We create its own rendering file which extends the Popover
renderer and specify its own property apiVersion (with value 2).

BCP: 2170166924
Change-Id: I7ee8f389f54a51a7b7756e80f9c4129e04617551