[FIX]P13nDialog: improve scrolling behaviour for mobile devices
If only one panel is provided for the P13nDialog,
the P13nDialog#_switchPanel function is never executed.

The device check for the initial propagation of the 'verticalScrolling'
prevents the P13nDialog from displaying the scrollbar on any mobile device
in case only one panel has been provided. Usually this is not an issue,
as there are several panels in the P13nDialog (thus P13nDialog#_switchPanel
always gets executed).

Removing this check will propagate the scrolling config from the panel
as expected in case only one panel is visible and a mobile device is used.

BCP: 002075129500003528992021
Change-Id: I06f0ffa341309831d4b7eed4e30a3e9a80d91a73