[INTERNAL] sap.ui.core: allow using Core via AMD dependency
There was one last gap when using sap/ui/core/Core via an AMD
dependency: When a custom boot task ('xx-bootTask') was used and when
that task synchronously required a module depending on the Core, then a
dependency cycle was formed and the module received an 'undefined' value
for the Core dependency.

By moving the execution of the custom boot task and all follow-up steps
from the Core constructor to Core's boot() method, the Core module is
fully executed before the custom boot task starts executing.

Still, a custom boot task should be able to do the same changes to the
Core configuration as before.

Potential incompatibilities:
- the custom boot task can no longer intercept Core.boot() as it isn't
  executed early enough. This never was a supported or suggested
  scenario, no such usages are known and custom boot tasks are not a
  public offering
- code that requires sap/ui/core/Core early, no longer can expect
  the moved code to be executed. Again, this is not a documented usage
  and all known bootstrap code calls Core.boot() immediately after
  requiring it.
- tests that observe the order of sap.ui.requireSync calls during
  bootstrap will notice a different order than before as the Core facade
  is created at an earlier point in time.
  The only known affected test (syncXHRBootstrap) has been adapted

Change-Id: I597f9bd0192443270b99482732ff64c4cdf2a384