[INTERNAL] RenderManager: Enable running without inline styles for CSP
 - This change introduces three important aspects to the rendering.

1. Avoids writing inline styles to the DOM
 - In string-based rendering, the RM#writeStyles method is not writing
styles into the HTML buffer any longer, instead stores the styles in the
memory and adds a style-key marker that is referencing to the original
styles. After the control rendering, including descendants, has been
finalized and before the HTML buffer is injected to the DOM, the HTML
buffer is converted to the DOM nodes, afterwards the original styles are
set via DOM CSSStyleDeclaration API. Converting HTML buffer to DOM nodes
in the initial string-based rendering helps us not to touch DOM
afterwards for styles. Also using the
CSSStyleDeclaration API ensures not to break out of the style
attribute.

2. Patcher no longer requires style-src: 'inline-style'
 - Previously the Patcher was using the setAttribute("style") API to set
styles for DOM nodes. In this version the Patcher also uses
CSSStyleDeclaration API to update and set styles to ensure CSP
compliance. To increase the security, Patcher.style API ensures a single
style can be added per API call. Also Patcher now supports using the
native CSS.escape API. For internal use only,
Patcher.unsafeHtml method now accepts a callback parameter to process
the provided HTML blob as DOM nodes after it is injected into the DOM
tree.

3. DOM-based initial rendering
 - Previously the initial rendering was always string-based rendering
even though the root control supports DOM APIs via apiVersion:2 marker.
As explained in the bullet 1 and 2, converting HTML buffer to DOM nodes,
storing styles in the memory afterwards searching the style-key
attribute in the DOM tree and finally setting the styles will not be
necessary if we start building the DOM tree structure via the DOM API
for the controls using the apiVersion:2 marker. For the sake of initial
rendering performance, this change now introduces DOM-based initial
rendering when possible. This will help us to simplify the rendering.

Jira: FIORITECHP1-20763
Jira: CPOUIFTEAMB-1820

Change-Id: Ie27057ab7bb5291a7551a5a2c073f9908544ad73