[FIX] sap.m.FacetFilter: has stable tests
Issue: one of the test cases checks if the FacetFilterList is contained
in the correct aggregation after closure of the Popover used to show the
FacetFilter. On touch devices there is an animation while closing, which
was often causing failiures of the test.

Solution: the test is timeouting the assertion in order to wait
for the animation to be over.

BCP: 2170145118
Change-Id: If83eee46cb0073ce63eef3f1408a38e0c09456b0