[INTERNAL][FIX] Table: Clear selection on binding length change
If the binding length changes, e.g. on expand/collapse, the indices that are
selected in the SelectionModel are invalid. To avoid an incorrect selection
state with the experimental grouping and with ODataV4 grouping/aggregation, the
selection is cleared just like it is already done on sort and filter, for
example.
The fix is done in the SelectionModelSelection plugin, which is automatically
applied to instances of sap.ui.table.Table if no selection plugin is applied
externally. The MultiSelectionPlugin internally uses the SelectionModelSelection
plugin for instances of sap.ui.table.Table.
No change for TreeTable and AnalyticalTable required. They use the
BindingSelection plugin, where the selection is maintained by the bindings, who
are able to handle this case.

Adds QUnit module for the SelectionModelSelection plugin with some initial
tests.

Change-Id: I6590b5402c8b317374543bae63f6ea6f08183803