[INTERNAL] sap.f.GridContainer: Make the grid more stable in context of a CSSGrid
When the GridContainer was invalidated in context of a CSSGrid there was a flickering.
Now this is fixed by executing logic for enforce max columns only after resize,
and not during the rendering and setting of styles.

Also columnsChange event is fired only after resize and not during initial rendering.

Change-Id: Iabf189f76180b0868c0eacd8cb4c4245621b70a3
BCP: 2170117632