[FIX] sap.m.MessagePopover: Update internal item properties when an external is changed
When having a single item bound and a new model with same amount of items is set,
the framework is just forwarding the properties to the item without informing
that aggregation has been changed. Message Popover did not listen for such changes
and used to have the old items populated inside.

Change-Id: I6c134aad644cd996ec7f8c7799006ecd700d2392
BCP: 2170113391
CR-Id: 002075125900002750502021