[FIX] sap.ui.integration.cards.AdaptiveContent: Refresh CSS properly on theme changed
When the theme changes, we need to update the CSS variables. So, far
they were cached in the hostConfig and changing the theme was causing
strange behaviour where some of the stylings were applied, but not all.

With current implementation, the hostConfig is not a static JSON
anymore as it should react and refresh dynamically and on demand.

BCP: 2180176493
Fixes: https://github.com/SAP/openui5/issues/3261

Change-Id: If3f616bf8682a2029efc5e44727c7ec5378ac12e