[INTERNAL] sap.m.Input: Prevent Dialog rerendering when show all items buttton gets visible
When adding the end button aggregation to the dialog, the dialog gets
rerendered which causes rerendering of its content. That replaces the input's
DOM and it loses focus. When keyboard layout is switched (e.g. we are on the
numbers keyboard layout) and focus is lost, the keyboard switches to the
default view (which is usually letters). That issue is fixed with semantic
rendering implementation and the fix is specific for 1.71.

Change-Id: I4f6742e857535fb4cc88eb6db4a24c309040e1a4
BCP: 002075129400003135412021
CR-Id: 002075125900002404592021