[FIX] sap.m.Popover: Race condition during openning fixed
There are cases in which the rersize handler calls _fnOrientationChange -> _applyPosition
before Popover's animation starts. This is unexpected behaviour as there's no need for those
calculations- the Popover would do them after it's been openned, and this could lead to side effects
with wrongly callculated positions.
Skipping/Suppressing ResizeHandler's callback if the Popover is not entirely openned would solve the
issue. When the Popover is finally open, there are no other calculations than the ones
in the ResizeHandler.

This race condition is reproducible in NavContainer's "Focus is changed only after transition" test.

Change-Id: Ia13bd78fc9422b18cfd2261d9483cf4453ca8e53
BCP: 2180206885