[INTERNAL][FIX]AdaptFiltersPanel: 'visibleInDialog' as model Filter
The 'visibleInDialog' attribute should be handled as model filter
instead of manipulating the visibility through a binding.

A Bug caused by the prior implementation was that the growing text was
not correctly announced for the hidden items.

(The tests have been refactored to use 'prepareAdaptationData' instead
of the deprecated 'prepareP13nData' model creation)

BCP: 2170144520
Change-Id: I1ab1f0bc35a1e9ab1322393e393734184349790b