[INTERNAL] sap.m.SelectList: Adaptation to OData v4
Problem: In some cases, unwanted PATCH request may be triggered from
Select control, when using data binding with OData v4.
This happens when "forceSelection" property is set to "true" (default
value). When updateItems is called with AddVritualContext, the Select
control executes its synchronizeSelection logic - the method is called
once synchronously and second time asynchronously. In the async call,
when "forceSelection" is "true", the "selectedKey" property is being
updated. However, before the real items updated, there is only one
(dummy) item in the "items" aggregation, which is used as s template for
the data binding (and is added with AddVirtualContext reason). This
item does not have any properties set and its "key" property is an empty
string. If the "selectedKey" of the Select control is binded, at this
moment, there is an unwanted PATCH request sent for updating with a null
value.

Solution: Do not try to synchronize the selection, when updateItems is
called with AddVirtualContext and RemoveVirtualContext reasons. Only
aggregation should be updated, any additional logic of the control
should be ignored.
With OData v2, there is no updateItems calls with Add/Remove
VirtualContext. udpateItems is only called when items are in real
updated.

JIRA: BGSOFUIPIRIN-5148
Change-Id: I08f691014f2b85a6a5846db5016bca563345b3a8