[INTERNAL][FIX] sap.m.Table: _getInitialAccumulatedWidth returns wrong value
The mentioned function returns a wrong value of the fAccumulatedWidth when the effected
sap.m.Table is placed within a sap.m.IconTabBar. On tab switch the table is not yet presented
inside the DOM element therefore $this.closest(".sapUiSizeCompact").length return an empty array
that results to a bad calculation of the columns width and minScreenWidth.

BCP: 2180172830
Change-Id: Ib6a73bd18ec46ee0cb5330e5ea3e80e42f0665b4