[INTERNAL] core.LocaleData: remove M_DEFAULT_DATA, load en.json (CLDR) instead
Remove duplicated data from LocaleData.js:
* M_DEFAULT_DATA is a reduced copy of sap/ui/core/cldr/en.json

If locale is unknown, load en.json instead of M_DEFAULT_DATA.

Introduces LocaleData#firstDayStartsFirstWeek to determine how
the calendar week is calculated (US).
true means that the first day of the year (January the first)
is always the first day of the first calendar week of the year.

It can be used to calculate US calendar week numbers.
E.g. for year 2021
01.01. => CW 1
31.12. => CW 53

Change-Id: I14b4c67fb9db3dd6c228bb21578d7798d648cc0c