[INTERNAL][FIX]UIManager: improve escape handling
Downport for 5291579.

The 'AdaptFiltersDialog' in MDC is using the UIManager for the Dialog
creation. The UIManager in some cases did not correctly provide the
escape handler, as a consequence the escape handler has been changed
to always properly cancel the Dialog.

As this change is categorized as regression for accessibilty product
standards, a downport is required.

BCP: 2180177315
CR-Id: 002075125900002514592021
Change-Id: Ib27046d6366c7719fa70f12e67ea49d4e6394106