[FIX] sap.f.DynamicPageTitle: Removed potential redundant tab stops
Problem: When we have an sap.m.Carousel as a snappedContent aggregation
for the sap.f.DynamicPageTitle and the snappedContent is currently
hidden, we have tab stops in the content of the sap.m.Carousel.

Root Cause: The sap.m.Carousel CSS overwrites the "visibility" property
inherited from the sap.f.DynamicPageTitle. From "hidden" becomes "visible".

Solution: We add a new CSS rule which forces the snappedContent and
its children to be with "visibility" hidden when the snappedContent is
supposed to be.

BCP: 2180200701
Change-Id: I345a6a1a7e643389372f152d11b3ce54984238c9