[FIX] sap.m.DatePicker: selects correct date
Issue: When the user picks a month from a DatePicker with dispalyFormat
with only month and year and the new month has fewer dates from the one
previously selected, the result date saved in the control was wrong.

Example: The initial selected date is 31 May, the user selects September
and the result was October. After this change, the correct result is
September.

BCP: 002075129500003918772021
Change-Id: I6e78d938a255ec494528c8263d13ec61ebb9496a