[FIX] sap.m.Popover: Calculate properly dimensions when the within area is bigger that the screen
There are edge cases whete the within area is bigger than the screen or
there's additional margin out of the UI5 app area. For such scenarios
we need to make the calculations based on the scroll area of the body
instead of window dimensions. In most of the cases they'd be the same.

Change-Id: Ie420b921d37a15669cd7e214e533760b229b3a17
BCP: 2170132760