[FIX] sap.f.GridContainer: Improve focus handling when using ComboBox inside the grid
Solving the following problem:
When user selects value from a ComboBox with mouse,
the grid navigation was deciding to force to
move the focus on the corresponding item wrapper.
Which was preventing the selection.

Now this is resolved with additional check
for if the active element is in the active item.
If so the focus will not be moved.

Change-Id: Id609820f7e514f4336f539176ad9968ffc08d78d
BCP: 2180211551