[INTERNAL] sap.ui.fl: move parsing of revert data to an earlier point
previously the revert data saved in the custom data of the control were
only parsed and saved in the change instance when the revert was triggered.
But the revertData are also necessary for other things (e.g. ChangeViz, Condenser, ...)
so the parsing happens now immediately in the Applier, when we see that the
change is already applied on the control, but the change instance has the wrong state.

This is a critical fix as the missing revert data can break the Condenser
when new changes are created.
If a change handler doesn't properly safeguard the revert data access, this
will prevent the creation of any further changes of this type and Save & Exit
will fail with a cryptic error message.
Currently this is known to affect the MoveControls change handler, which is
frequently used by customers.

BCP: 2180214137
CR-Id: 002075125900003469122021
Change-Id: Ibd81879731f16d302ea928b91c29a9b3153bc077
(cherry picked from commit a7fafca924e08a356dac99ef7feacfa2386cb0c9)