[INTERNAL] sap.ui.fl: move parsing of revert data to an earlier point
previously the revert data saved in the custom data of the control were
only parsed and saved in the change instance when the revert was triggered.
But the revertData are also necessary for other things (e.g. ChangeViz, Condenser, ...)
so the parsing happens now immediately in the Applier, when we see that the
change is already applied on the control, but the change instance has the wrong state.

BCP: 2180214137
Change-Id: Ibd81879731f16d302ea928b91c29a9b3153bc077