[INTERNAL] sap.f.DynamicPage: Fixed footer overlap with content
Problem: In some cases when we have both CSS classes
"sapFDynamicPageContentFitContainerFooterVisible" and
"sapFDynamicPageContentFitContainer" the content of the DynamicPage
overlaps with its footer.

Problem: We are relying on a pseudo element to make sure that there
is enough space underneath the footer and to eliminate overlap but it
seems it's not always working.

Solution: Now in the problematic case, we force the maximum height of
of the DynamicPage's content, to be its original height subtracted by
the height of the footer.

BCP: 2180195326
BCP: 2180208786
Change-Id: I660bf8db65b700504b520cf0df9ff1f2c1220206