[FIX] sap.f.DynamicPage: Fix pin button visibility
Problem: When DynamicPage is placed in FlexibleColumnLayout and the
closes the second column, pin button appears, despite the fact that
preserveHeaderStateOnScroll is "true". The problem happens, when the
DynamicHeader occupies more than 60% of the screen height (when second
column is shown, the first one is smaller) and after that second column
is closed and the first one gets bigger (so that the DynamicHeader now
occupies less than 60% of the screen). It happens due to a rerendering
of the DynamicPage, just before the resize of the columns of the
FlexibleColumnLayout. After the rerendering the "sapUiHidden" class of
the pin button is deleted. When onAfterRendering logic is executed (and
the logic for showing/hiding pin button is called), the first column of
the FlexibleColumnLayout is still not resized, so the pin button is
shown (as the DynamicPageHeader at this moment is bigger than 60% of the
page's height).

Solution: Now we update the pin button visibility after resizing of the
DynamicHeader too, so when it gets its final height, we can properly
update the button.

BCP: 002075129500003206892021
Change-Id: I5e2856b2cf3eeaf8400ca385a557738d3cf759f3