[FEATURE] mdc/Table: enableAutoColumnWidth property added
 - This feature provides a heuristic column size handling for the
mdc/Table based on metadata information.
 - PropertyHelper#setColumnWidth method is added to calculate the width
from the type, label, constraints,... of the PropertyInfo
 - When a new Column is added now the Table calculates and sets the
column width if the PropertyHelper is already available, if not it waits
until it is resolved.
 - To influence the width calculation algorithm
visualSettings.widthCalculation key
of the PropertyInfo can be used while creating PropertyInfos.

 - Sample pages are adapted to reflect the new column width algorithm.
 - Metadata of the Products xml file is changed accordingly.
 - Sample TableDelegate now passes maxLength, precision, scale
constraints to the type util for better column width results.
 - UnitTests are adapted even though the code coverage is in an
acceptable state there should be further tests regarding column
calculation but since this calculation will move to the more central
place, here we do not want to waste too much time for the column width
calculation tests.

Jira: CPOUIFTEAMB-1624
Change-Id: I82744fcc1f8ed6e80881b451e0fd117e85b8b9e6