[INTERNAL] sap.ui.fl.registry.Settings offers isPublicFlVariantEnabled
To determine if the fl-VariantManagement should show a "Public" button, the capability must be given by the back end as well as the user authorized to do so.
Thus the settings can be provided by the back end or can be determined by the client in case the back end does not (yet) provide this flag.

Change-Id: Ibffe871a4d17d6714470902c01e26225f3ae6d60
JIRA: CLOUDIOI-444