[INTERNAL] Table: update trigger button width when column gets hidden via column resize
When the Table is configured with hiddenInPopin, the columns are hidden via
setDisplay method of the sap.m.Column. This function hides the column by changing
the DOM appropriately and the Table control does not rerender. Due to this,
when there is Dummy column rendered and growing trigger button is visible
in the table, the width of the growing trigger button is not updated as expected
when column gets hidden. This is now fixed by updating the growing trigger button
width via onColumnResize method of the table. The GrowingEnablement is also
refactored where the a method is introduced to update the trigger width.
BCP: 2180167536

Change-Id: I3c94ab70b7a6f8eacb1bb57d793861d615975ae4