[INTERNAL] ListBase: fixes and update documentation for requestItems
The requestItems method of the ListBase is updated to clarify its
behavior when its parameter is not specified. In this case the
growingThreshold value is taken into consideration for requesting more
data for the list.
Only positive numbers are allowed as parameter for this API.
If a negative number of 0 is passed as a parameter or growing is
disabled for the list control, then an Error is thrown.

Change-Id: I930564b01321f82569f73ccbb7dfd5c40b18434d