[FEATURE] Field/FilterField: support field help for external content
If a content control has a "showValueHelp" property, a
"valueHelpIconSrc" property and a
"valueHelpRequest" event (like sap.m.Input) the field help of the
Field/FilterField is also working for this content control.

The content control might need to set field help specific
aria-attributes by it's own as the Field/FilterField cannot know what
needs to be suppored inside of this content control.

Change-Id: I1672d76b09f01832265b13c0887fb3bedef58fc0
JIRA: CPOUIFDENVER-1650