[INTERNAL] ESLint: Activate more rules
Trivial findings are fixed within this change.
More complex findings are ignored via a eslint-disable comment and should
be addressed as follow-up by the respective code-owners.

Some findings could have been fixed, but the main focus of this change is
to enable the new rules without risking to change the runtime behavior
unexpectedly.

Change-Id: Ie5fad2cb5f19973db4caa93fe5eb7009a0ba476a
JIRA: CPOUI5FOUNDATION-373