[INTERNAL] ui5loader: make asyncMode test more robust
When the "Async/Sync Conflict : Warning Message" test (A) was executed
immediately before the „Unmanaged Module Definitions : Unnamed
Definition“ test (B), then it could happen, that the module created by
(B) was processed in the "onload" event of (A). In that case, (B) got
not only assigned the name of (A) but also the expected error was not
thrown and rather rejected the promise for (B).

The test has been fixed by waiting in (A) for the "onload" event of the
obsolete async request before continuing with the next test.

Note: The fact that scripted modules might be processed at a wrong /
unexpected point in time (here: "onload" event of (A)), is hard to
avoid. There's already a warning that they must be given a name to avoid
issues. The only feasible alternative would be to throw an error
instead, based on document.currentScript and aExecStack, but this might
cause regressions in existing code and therefore was ruled out.

Change-Id: I227b5c301f43cf103fe9d5f0f24599bfa28e8097