[INTERNAL] Integration Cards: Isolate the card from a default model
Solving the following issue:
If a container which holds the card has a default model,
it is propagated to the card and some features break.

Now the card always has a default model, so that
the propagation does not happen.
In the documentation of the managed object,
it can be seen that this is the recommended way to isolate.

Change-Id: I5f81ef40ac195af73fa1e68f0d45a11072c6e267
Snippix: 40537
BCP: 2180259123