[FIX] sap.ui.core.InvisibleMessage: Span content is now cleared
In order to avoid reading the span content of the Invisible Message
when the Screen Reader is in reading mode, we should clear the span
content after a while.
Also, we check if there is a new announcement call in order to avoid clearing
the span before the announcement is made.

Change-Id: I16b43fdf4ade3570abc8e2fb3d5a0e3d2ba2284f
BCP: 2180195538
BCP: 002075129400002780742022
CR-Id: 002075125800002313502022